// src/app/shared/interfaces/indicadores-frota.model.ts
export interface IndicadoresFrotaMV {
  desvioKm?: {
    data: any[];
    filter: any[]
  };
  kmMedio?: {
    data: any[];
    filter: any[]
  };
  kmMedioUltimaRevisao?: {
    data: any[];
    filter: any[]
  };
  frotaAlugadaTempo?: {
    data: any[];
    filter: any[]
  };
  distribuicao?: {
    data: any[];
    filter: any[]
  };
  grupoVeiculo?: {
    data: any[];
    filter: any[]
  };
  modeloVeiculo?: {
    data: any[];
    filter: any[]
  };
  vencimentario?: {
    data: any[];
    filter: any[]
  };
  idadeMedia?: {
    data: any[];
    filter: any[]
  };
  kmMedioUf?: {
    data: any[];
    filter: any[]
  };
  kmMedioEstado?: {
    data: any[];
    filter: any[]
  };
}